Apply the necessary JavaScript to make the game functional
Added the JavaScript code that will add player's symbol to the cell
they click on the board. Player turns will now alternate between one
another and if the player clicks an already occupied cell, it will
wait for the player to make a valid move. The winner of the match
(or tie) will display at the end of the game.